peteLib now processes floats

+++ peteLib.h
@@ -1,3 +1,3 @@
 // Header file for peteLib
 
-extern "C" void add(int[], int[], int[], int);
+extern "C" void add(float[], float[], float[], int);
